+++ pandapower/run.py
@@ -7,7 +7,8 @@
 import numpy as np
 
 from pandapower.auxiliary import _add_pf_options, _add_ppc_options, _add_opf_options, \
-    _check_if_numba_is_installed, _check_bus_index_and_print_warning_if_high
+    _check_if_numba_is_installed, _check_bus_index_and_print_warning_if_high, \
+    _check_gen_index_and_print_warning_if_high
 from pandapower.optimal_powerflow import _optimal_powerflow
 from pandapower.opf.validate_opf_input import _check_necessary_opf_parameters
 from pandapower.powerflow import _powerflow
@@ -77,7 +78,8 @@ def _passed_runpp_parameters(local_parameters):
         'tolerance_kva': 1e-05,
         'trafo_loading': 'current',
         'trafo_model': 't',
-        'voltage_depend_loads': True
+        'voltage_depend_loads': True,
+        "trafo3w_losses": "hv"
     }
 
     passed_parameters = {
@@ -90,7 +92,7 @@ def _passed_runpp_parameters(local_parameters):
 def runpp(net, algorithm='nr', calculate_voltage_angles="auto", init="auto", max_iteration="auto",
           tolerance_kva=1e-5, trafo_model="t", trafo_loading="current", enforce_q_lims=False,
           numba=True, recycle=None, check_connectivity=True, r_switch=0.0, voltage_depend_loads=True,
-          delta_q=0, **kwargs):
+          delta_q=0, trafo3w_losses="hv", **kwargs):
     """
     Runs PANDAPOWER AC Flow
 
@@ -188,6 +190,8 @@ def runpp(net, algorithm='nr', calculate_voltage_angles="auto", init="auto", max
         **voltage_depend_loads** (bool, True) - consideration of voltage-dependent loads. If False, net.load.const_z_percent and net.load.const_i_percent are not considered, i.e. net.load.p_kw and net.load.q_kvar are considered as constant-power loads.
 
         **delta_q** - Reactive power tolerance for option "enforce_q_lims" in kvar - helps convergence in some cases.
+        
+        **trafo3w_losses** - defines where open loop losses of three-winding transformers are considered. Valid options are "hv", "mv", "lv" for HV/MV/LV side or "star" for the star point.
 
         ****kwargs** - options to use for PYPOWER.runpf
     """
@@ -232,18 +236,19 @@ def runpp(net, algorithm='nr', calculate_voltage_angles="auto", init="auto", max
         max_iteration = default_max_iteration[algorithm]
 
     # init options
-    # net.__internal_options = {}
     net._options = {}
     _add_ppc_options(net, calculate_voltage_angles=calculate_voltage_angles,
                      trafo_model=trafo_model, check_connectivity=check_connectivity,
                      mode=mode, copy_constraints_to_ppc=copy_constraints_to_ppc,
                      r_switch=r_switch, init=init, enforce_q_lims=enforce_q_lims,
-                     recycle=recycle, voltage_depend_loads=voltage_depend_loads, delta=delta_q)
+                     recycle=recycle, voltage_depend_loads=voltage_depend_loads, delta=delta_q,
+                     trafo3w_losses=trafo3w_losses)
     _add_pf_options(net, tolerance_kva=tolerance_kva, trafo_loading=trafo_loading,
                     numba=numba, ac=ac, algorithm=algorithm, max_iteration=max_iteration)
     # net.__internal_options.update(overrule_options)
     net._options.update(overrule_options)
     _check_bus_index_and_print_warning_if_high(net)
+    _check_gen_index_and_print_warning_if_high(net)
     _powerflow(net, **kwargs)
 
 
@@ -310,6 +315,7 @@ def rundcpp(net, trafo_model="t", trafo_loading="current", recycle=None, check_c
     _add_pf_options(net, tolerance_kva=tolerance_kva, trafo_loading=trafo_loading,
                     numba=numba, ac=ac, algorithm=algorithm, max_iteration=max_iteration)
     _check_bus_index_and_print_warning_if_high(net)
+    _check_gen_index_and_print_warning_if_high(net)
     _powerflow(net, **kwargs)
 
 
@@ -385,6 +391,7 @@ def runopp(net, verbose=False, calculate_voltage_angles=False, check_connectivit
                      voltage_depend_loads=False, delta=delta)
     _add_opf_options(net, trafo_loading=trafo_loading, ac=ac, numba=numba)
     _check_bus_index_and_print_warning_if_high(net)
+    _check_gen_index_and_print_warning_if_high(net)
     _optimal_powerflow(net, verbose, suppress_warnings, **kwargs)
 
 
@@ -446,4 +453,5 @@ def rundcopp(net, verbose=False, check_connectivity=True, suppress_warnings=True
                      voltage_depend_loads=False, delta=delta)
     _add_opf_options(net, trafo_loading=trafo_loading, ac=ac)
     _check_bus_index_and_print_warning_if_high(net)
+    _check_gen_index_and_print_warning_if_high(net)
     _optimal_powerflow(net, verbose, suppress_warnings, **kwargs)